ai/requirements.txt
google-genai
fastapi
pydantic
uvicorn[standard]